src/scripts/requirements.txt
# Playback Python Dependencies
# Required for recording, processing, and OCR functionality

# PyObjC framework bindings for macOS integration
pyobjc-framework-Vision>=10.0      # Apple Vision framework for OCR
pyobjc-framework-Quartz>=10.0      # Core Graphics and image loading
pyobjc-framework-Foundation>=10.0  # Foundation framework (NSURL, etc.)

# Testing (if needed)
# pytest>=7.0.0
# pytest-cov>=4.0.0

# Linting (if needed)
# flake8>=6.0.0
